WIP: the great refactoring
