+++ Lib/diffenator/__main__.py
@@ -41,26 +41,10 @@ Diff rendered glyphs:
 diffenator /path/to/font_a.ttf /path/to/font_b.ttf -td glyphs -rd
 """
 from argparse import RawTextHelpFormatter
-from diffenator.diff import diff_fonts
-from diffenator import __version__
-from diffenator.font import InputFont
-from diffenator.utils import (
-    vf_instance,
-    vf_instance_from_static
-)
-from diffenator.reporters import diff_report, CLIFormatter, MDFormatter
+from diffenator import CHOICES, __version__
+from diffenator.font import DFont
+from diffenator.diff import DiffFonts
 import argparse
-
-
-DIFF_CHOICES = [
-    'names',
-    'marks',
-    'mkmks',
-    'attribs',
-    'metrics',
-    'glyphs',
-    'kerns'
-]
 
 
 def main():
@@ -70,7 +54,7 @@ def main():
 
     parser.add_argument('font_a')
     parser.add_argument('font_b')
-    parser.add_argument('-td', '--to_diff', nargs='+', choices=DIFF_CHOICES,
+    parser.add_argument('-td', '--to_diff', nargs='+', choices=CHOICES,
                         default='*',
                         help="categories to diff. '*'' diffs everything")
 
@@ -81,8 +65,8 @@ def main():
     parser.add_argument('-v', '--verbose', action='store_true',
                         help='Report diffs which are the same')
 
-    parser.add_argument('-i', '--vf-instance', default='Regular',
-                        help='Variable font instance to diff')
+    parser.add_argument('-i', '--vf-instance', default='wght=400',
+                        help='Set vf variations e.g "wght=400"')
 
     parser.add_argument('--marks_thresh', type=int, default=0,
                         help="Ignore modified marks under this value")
@@ -92,52 +76,48 @@ def main():
                         help="Ignore modified kerns under this value")
     parser.add_argument('--glyphs_thresh', type=float, default=0,
                         help="Ignore modified glyphs under this value")
+    parser.add_argument('--metrics_thresh', type=int, default=0,
+                        help="Ignore modified metrics under this value")
     parser.add_argument('-rd', '--render_diffs', action='store_true',
                         help=("Render glyphs with hb-view and compare "
                               "pixel diffs."))
     parser.add_argument('-r', '--render-path',
-                        help="Path to generate before and after gifs.")
+                        help="Path to generate before and after gifs to.")
     args = parser.parse_args()
 
-    font_a = InputFont(args.font_a)
-    font_b = InputFont(args.font_b)
-
-    if 'fvar' in font_a and 'fvar' not in font_b:
-        font_a = vf_instance_from_static(font_a, font_b)
-
-    elif 'fvar' not in font_a and 'fvar' in font_b:
-        font_b = vf_instance_from_static(font_b, font_a)
-
-    elif 'fvar' in font_a and 'fvar' in font_b:
-        font_a = vf_instance(font_a, args.vf_instance)
-        font_b = vf_instance(font_b, args.vf_instance)
-
-    diff = diff_fonts(
-        font_a,
-        font_b,
-        categories_to_diff=args.to_diff,
-        glyph_threshold=args.glyphs_thresh,
-        marks_threshold=args.marks_thresh,
-        mkmks_threshold=args.mkmks_thresh,
-        kerns_threshold=args.kerns_thresh,
-        render_diffs=args.render_diffs
+    diff_options = dict(
+            marks_thresh=args.marks_thresh,
+            mkmks_thresh=args.mkmks_thresh,
+            kerns_thresh=args.kerns_thresh,
+            glyphs_thresh=args.glyphs_thresh,
+            metrics_thresh=args.metrics_thresh,
     )
+    font_a = DFont(args.font_a)
+    font_b = DFont(args.font_b)
 
-    report_formatter = CLIFormatter if not args.markdown else MDFormatter
-    report = diff_report(
-        diff,
-        args.font_a,
-        args.font_b,
-        Formatter=report_formatter,
-        output_lines=args.output_lines,
-        verbose=args.verbose
-    )
-    print(report)
+    if font_a.is_variable and not font_b.is_variable:
+        font_a.set_variations_from_static(font_b)
+
+    elif not font_a.is_variable and font_b.is_variable:
+        font_b.set_variations_from_static(font_a)
+
+    elif font_a.is_variable and font_b.is_variable:
+        variations = {s.split('=')[0]: float(s.split('=')[1]) for s
+                      in args.vf_instance.split(", ")}
+        font_a.set_variations(variations)
+        font_b.set_variations(variations)
+
+    diff = DiffFonts(font_a, font_b, diff_options)
+
+    if args.markdown:
+        print(diff.to_md(args.output_lines))
+    else:
+        print(diff.to_txt(args.output_lines))
 
     if args.render_path:
-        from diffenator.visualize import diff_render
-        diff_render(font_a, font_b, diff, args.render_path)
+        diff.to_gifs(args.render_path)
 
 
 if __name__ == '__main__':
     main()
+
